Avoid leaking capnp.Client in pkg/util/stream. Add BBR to streams.
To avoid leaking clients, we no longer check for failed calls at
the Stream.Call() call-site.  This greatly increases the risk of
OOM, since a slow consumer on the remote side can cause the len
of the queue to surge.

To mitigate the risks, we have added BBR flow control to existing
users of Stream.  We have also added appropriate warnings and
guidance in the package documentation.